backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
bcrypt==4.2.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
